+++ src/rules/armTemplate.ts
@@ -11,15 +11,10 @@ import {
   isRequestEvaluation,
   isAndEvaluation,
   HttpMethods,
+  filterAsync,
 } from '.';
 import {AzureIdentityCredentialAdapter} from '../azure';
 import {ScanResult} from '../scanner';
-
-// // needed for sendRequest method
-// // from @azure/core-http => https://azuresdkdocs.blob.core.windows.net/$web/javascript/azure-core-http/1.2.4/globals.html#httpmethods
-// enum HttpMethods {
-//   POST = 'POST',
-// }
 
 export interface ARMTemplate {
   $schema: string;
@@ -43,55 +38,6 @@ export interface ARMTarget {
   credential: TokenCredential;
   client: ResourceManagementClient;
 }
-
-// All evaluations contain a JMESPath query that operate on ARM resources
-// type BaseEvaluation = {
-//   query: string;
-// };
-
-// Some evaluations may check for additional conditions
-// type AndEvaluation = BaseEvaluation & {
-//   and: Array<Evaluation>;
-// };
-
-// type RequestEvaluation = BaseEvaluation & {
-//   request: {
-//     operation: string;
-//     query: string;
-//   };
-// };
-
-// function isAndEvaluation(evaluation: Evaluation): evaluation is AndEvaluation {
-//   return (evaluation as AndEvaluation).and !== undefined;
-// }
-
-// function isRequestEvaluation(
-//   evaluation: Evaluation
-// ): evaluation is RequestEvaluation {
-//   return (evaluation as RequestEvaluation).request !== undefined;
-// }
-
-// returns a resolved promise so that any async calls in the callback are completed before returning
-function mapAsync<T1, T2>(
-  array: T1[],
-  callback: (value: T1, index: number, array: T1[]) => Promise<T2>
-): Promise<T2[]> {
-  return Promise.all(array.map(callback));
-}
-
-// if the array is empty, it returns an empty array
-export async function filterAsync<T>(
-  array: T[],
-  callback: (value: T, index: number, array: T[]) => Promise<boolean>
-): Promise<T[]> {
-  // creates boolean array and maintains the same index order as the original array
-  const mappedArray = await mapAsync(array, callback);
-  // filters over the original array based on the mappedArray boolean at each index
-  return array.filter((_, index) => mappedArray[index]);
-}
-
-// Evaluations may be standalone or composite
-// export type Evaluation = BaseEvaluation | AndEvaluation | RequestEvaluation;
 
 export class ARMTemplateRule implements BaseRule<ARMTarget> {
   type: RuleType.ARM;
